# back/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OpenApi documentación de Minibares",
        "description": "Class LogController",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles_permisos": {
            "get": {
                "tags": [
                    "RolesPermisos"
                ],
                "summary": "List all roles_permisos",
                "description": "Display a listing of the resource.",
                "operationId": "245c38cfaa7de1d7ca7e1b7de0b58f56",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                }
            }
        },
        "/api/roles_permisos/{roleId}": {
            "post": {
                "tags": [
                    "RolesPermisos"
                ],
                "summary": "Registra un permiso de un rol",
                "description": "Registra un permiso de un rol",
                "operationId": "storeRolesPermisos",
                "parameters": [
                    {
                        "name": "roleId",
                        "in": "path",
                        "description": "ID del rol",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roleId": {
                                        "description": "Role Id relacionado",
                                        "type": "string"
                                    },
                                    "controlador": {
                                        "description": "Controlador que se ejecuta",
                                        "type": "string"
                                    },
                                    "accion": {
                                        "description": "Acción que se ejecuta",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Permiso creado correctamente"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/roles_permisos/{id}": {
            "get": {
                "tags": [
                    "RolesPermisos"
                ],
                "summary": "Obtiene un permiso de un rol",
                "description": "Obtiene un permiso con el id especificado de un rol",
                "operationId": "showRolesPermisosById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del permiso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "RolesPermisos"
                ],
                "summary": "Actualiza un permiso de un rol",
                "description": "Actualiza un permiso con el id especificado de un rol",
                "operationId": "updateRolesPermisosById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del permiso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "roleId": {
                                        "description": "Role Id relacionado",
                                        "type": "string"
                                    },
                                    "controlador": {
                                        "description": "Controlador que se ejecuta",
                                        "type": "string"
                                    },
                                    "accion": {
                                        "description": "Acción que se ejecuta",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/roles_permisos{id}": {
            "delete": {
                "tags": [
                    "RolesPermisos"
                ],
                "summary": "Elimina un permiso de un rol",
                "description": "Elimina un permiso con el id especificado de un rol",
                "operationId": "destroyRolesPermisosById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del permiso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
...
    }
}